refactor(schema): move value helpers to avro utils
HoodieSchemaUtils is the home of table-schema transforms, but it
carried two value-level helpers that only forwarded to HoodieAvroUtils:

- convertValueForSpecificDataTypes(HoodieSchema, Object, boolean) had
  no production callers (HoodieAvroUtils#getNestedFieldVal calls its own
  Schema-typed overload) and is deleted; its ten unit tests move to
  TestHoodieAvroUtils against the Avro overload with the assertions
  unchanged.
- convertBytesToBigDecimal(byte[], HoodieSchema) moves next to the
  other decimal decoders in HoodieAvroUtils and now calls the
  (byte[], int, int) overload directly; the HoodieSchemaUtils
  (byte[], int, int) shim is deleted. DecimalLogicalTypeProcessor and
  TestJsonKafkaSource switch to the new home.

TestDataSourceDefaults passes the raw Avro field schema instead of
wrapping it only to unwrap it again. Two tests pin the moved decimal
decoder (bytes-backed value; non-DECIMAL schema rejected).

Part of #16639.
